feat: Navbar 재생성
기존 Navbar 삭제 후 세분화한 Navbar로 재생성
